django_server/requirements.txt
Django~=5.1.2
django-oauth-toolkit~=3.0.1
requests~=2.32.3
Pillow~=11.0.0
influxdb-client~=1.46.0
django-environ~=0.11.2
djangorestframework~=3.15.2
django-cors-headers~=4.5.0
python-decouple~=3.8
channels~=4.1.0
asgiref~=3.8.1
daphne~=4.1.2
apscheduler~=3.10.4
opencv-python~=4.10.0.84
